采购入库 标签改为标准标签

+++ jeeplus-module/jeeplus-test/src/main/java/com/jeeplus/modules/business/shengchan/dingdan/web/BusinessShengChanDingDanController.java
@@ -184,7 +184,7 @@ public class BusinessShengChanDingDanController extends BaseController {
 	}
 	@ResponseBody
 	@RequestMapping("chaidan")
-	public AjaxJson chaidan(String rid,int num){
+	public AjaxJson chaidan(String rid,Double num){
 		AjaxJson json = new AjaxJson();
 		try {
 			businessShengChanDingDanService.chaidan(rid,num);
